+++ demisto_sdk/commands/lint/resources/pylint_plugins/certified_partner_level_checker.py
@@ -32,6 +32,16 @@ cert_partner_msg = {
         "main-func-doesnt-exist",
         "Please remove all prints from the code.",
     ),
+    "W9008": (
+        "Do not use demisto.results function. Please return CommandResults object instead.",
+        "demisto-results-exists",
+        "Do not use demisto.results function.",
+    ),
+    "W9009": (
+        "Do not use return_outputs function. Please return CommandResults object instead.",
+        "return-outputs-exists",
+        "Do not use return_outputs function.",
+    ),
     "W9016": (
         "Initialize of params was found outside of main function. Please use demisto.params() only inside main "
         "func",
@@ -67,6 +77,8 @@ class CertifiedPartnerChecker(BaseChecker):
 
     def visit_call(self, node):
         self._sys_exit_checker(node)
+        self._return_outputs_checker(node)
+        self._demisto_results_checker(node)
         self._init_params_checker(node)
         self._init_args_checker(node)
 
@@ -111,6 +123,36 @@ class CertifiedPartnerChecker(BaseChecker):
                 and node.args[0].value != 0
             ):
                 self.add_message("sys-exit-exists", node=node)
+
+        except Exception:
+            pass
+
+    def _return_outputs_checker(self, node):
+        """
+        Args: node which is a Call Node.
+        Check:
+        - if return_outputs() statement exists in the current node.
+
+        Adds the relevant error message using `add_message` function if one of the above exists.
+        """
+        try:
+            if node.func.name == "return_outputs":
+                self.add_message("return-outputs-exists", node=node)
+
+        except Exception:
+            pass
+
+    def _demisto_results_checker(self, node):
+        """
+        Args: node which is a Call Node.
+        Check:
+        - if demisto.results() statement exists in the current node.
+
+        Adds the relevant error message using `add_message` function if one of the above exists.
+        """
+        try:
+            if node.func.attrname == "results" and node.func.expr.name == "demisto":
+                self.add_message("demisto-results-exists", node=node)
 
         except Exception:
             pass
