+++ demisto_sdk/tests/test_files/py3_XSOARLinter_demisto_results_warnings.py
@@ -74,7 +74,6 @@ def parse_incidents(
 
 
 def fetch_incidents():
-    params = demisto.params()
     return []
 
 
@@ -97,16 +96,16 @@ def main():
         proxy=proxy,
     )
     command = demisto.command()
-    demisto.info(f"Command being called is {command}")
+    LOG(f"Command being called is {command}")
     # Commands dict
     commands: Dict[str, Callable[[Client, Dict[str, str]], Tuple[str, dict, dict]]] = {
         "test-module": test_module,
     }
     if command in commands:
-        return_results(*commands[command](client, demisto.args()))
+        return_outputs(*commands[command](client, demisto.args()))
     elif command == "fetch-incidents":
         incidents = fetch_incidents()
         demisto.incidents(incidents)
-        return_results(incidents)
+        demisto.results(incidents)
     else:
         raise NotImplementedError(f"{command} is not an existing QuestKace command")
